chore: remove workspace structure from aria-query repository
