--- crates/aria_query/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "aria_query"
-version = "0.1.0"
-authors.workspace = true
-categories.workspace = true
-description.workspace = true
-edition.workspace = true
-keywords.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-dom = { path = "../dom" }
