--- crates/aria_query/src/lib.rs
@@ -1,1 +1,0 @@
-pub use ::dom::*;
